Contributors and bug fix (#2)
* Update version

* Fix null id

* Update log property

* Update .gitignore

* request_limit is valid only for ebooks

* Tag people by username

* Add logs

* v2.0.0

* Fix error with forwared messages

* Update error messages

* Fix other tags error

* Manage requests without link

* Manage special tags

* Update message

* Fix message

* Update messages

* /done command

* Mark as done with file

* Add contributors help

* Get last date when requesting audiobooks

* Update audiobooks message

* Delete message when marking done manually

* Reply to message and improve help

* Add admin name

* Update log

* Me command

* Delete mark done with file messages

* Update log

* Fix bugs

* Remove done command better

* Add sdone command

* Fix allow

* Fix SuperAdmin commands

* Fix come back again message when 0 hours left

* Update Request UQ

* Fix replyTo

* Update repeted request date only when valid

* Add /pending

* Update conditions order

* First implementation

* /request gets only new requests

* Schedule delete alive responses

* Add cancel command

* Add /cancel as reply to message

* Remove duplicate

* Add /requests in PM

* Add group name

+++ src/main/java/com/pirasalbe/services/telegram/handlers/command/TelegramSuperAdminCommandHandlerService.java
@@ -109,22 +109,22 @@ public class TelegramSuperAdminCommandHandlerService {
 		// prepare keyboard
 		InlineKeyboardMarkup keyboard = new InlineKeyboardMarkup();
 		sendMessage.replyMarkup(keyboard);
-		pagination.getElements().forEach(
-				admin -> keyboard.addRow(new InlineKeyboardButton(admin.getId() + " - " + admin.getRole().name())
+		pagination.getElements().forEach(admin -> keyboard.addRow(
+				new InlineKeyboardButton(admin.getName() + " - (" + admin.getId() + ") - " + admin.getRole().name())
 						.callbackData(COMMAND_COPY + " " + admin.getId())));
 
 		// add navigation buttons
 		List<InlineKeyboardButton> navigationButtons = new ArrayList<>();
 		if (page > 0) {
 			InlineKeyboardButton previous = new InlineKeyboardButton("Previous page");
-			int newOffset = page - size >= 0 ? page - size : 0;
-			previous.callbackData(COMMAND_LIST + " " + newOffset + " " + size);
+			int newPage = page - 1;
+			previous.callbackData(COMMAND_LIST + " " + newPage + " " + size);
 			navigationButtons.add(previous);
 		}
 		if (page < pagination.getTotalPages() - 1) {
 			InlineKeyboardButton next = new InlineKeyboardButton("Next page");
-			int newOffset = page + size;
-			next.callbackData(COMMAND_LIST + " " + newOffset + " " + size);
+			int newPage = page + 1;
+			next.callbackData(COMMAND_LIST + " " + newPage + " " + size);
 			navigationButtons.add(next);
 		}
 		if (!navigationButtons.isEmpty()) {
@@ -139,11 +139,10 @@ public class TelegramSuperAdminCommandHandlerService {
 			// ask the id
 			StringBuilder builder = new StringBuilder(COMMAND_ADD).append("\n");
 			builder.append("Send the ID of the user and the role").append("\n");
-			String placeholder = "id " + UserRole.getRoles();
-			builder.append("Format: <code>").append(placeholder).append("</code>");
+			builder.append("Format: id name <code>").append(UserRole.getRoles()).append("</code>");
 			SendMessage sendMessage = new SendMessage(TelegramUtils.getChatId(update), builder.toString());
 			sendMessage.parseMode(ParseMode.HTML);
-			sendMessage.replyMarkup(new ForceReply().inputFieldPlaceholder(placeholder));
+			sendMessage.replyMarkup(new ForceReply().inputFieldPlaceholder("id name " + UserRole.getRoles()));
 
 			bot.execute(sendMessage);
 		};
@@ -163,12 +162,18 @@ public class TelegramSuperAdminCommandHandlerService {
 		SendMessage sendMessage = null;
 
 		String[] parts = text.trim().split(" ");
-		if (parts.length == 2) {
+		if (parts.length == 3) {
 			// add user
 			long adminId = Long.parseLong(parts[0]);
-			UserRole adminRole = UserRole.valueOf(parts[1].toUpperCase());
-			adminService.insertUpdate(adminId, adminRole);
-			sendMessage = new SendMessage(chatId, "Admin " + adminId + " with role " + adminRole.name() + " added");
+			if (chatId != adminId) {
+				String name = parts[1];
+				UserRole adminRole = UserRole.valueOf(parts[2].toUpperCase());
+				adminService.insertUpdate(adminId, name, adminRole);
+				sendMessage = new SendMessage(chatId,
+						"Admin " + name + " (" + adminId + ") with role " + adminRole.name() + " added");
+			} else {
+				sendMessage = new SendMessage(chatId, "You can't add/edit your user");
+			}
 		} else {
 			sendMessage = new SendMessage(chatId, "There is something wrong with your message. Try again.");
 		}
@@ -208,8 +213,12 @@ public class TelegramSuperAdminCommandHandlerService {
 
 		// add user
 		long adminId = Long.parseLong(text.trim());
-		adminService.deleteIfExists(adminId);
-		sendMessage = new SendMessage(chatId, "Admin " + adminId + " removed");
+		if (chatId != adminId) {
+			adminService.deleteIfExists(adminId);
+			sendMessage = new SendMessage(chatId, "Admin " + adminId + " removed");
+		} else {
+			sendMessage = new SendMessage(chatId, "You can't remove your user");
+		}
 
 		sendMessage.replyToMessageId(messageId);
